% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cases.R
\name{get_cases}
\alias{get_cases}
\title{Retrieve Corona Virust Cases}
\usage{
get_cases()
}
\value{
a tidy dataframe of corona virus cases (infected, dead, cured)
}
\description{
This function returns a dataframe of corona virus cases. It takes no input.
}
\examples{
\dontrun{
cases <- refresh_cases()
}

}
