% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cases.R
\name{get_cases}
\alias{get_cases}
\title{Retrieve Corona Virus Cases}
\usage{
get_cases(wide = FALSE)
}
\arguments{
\item{wide}{A boolean expression to indicate if a wide dataframe is preferred. Default is FALSE.}
}
\value{
a tidy dataframe of corona virus cases (infected, dead, cured)
}
\description{
This function returns a dataframe of corona virus cases. It has an optional input of <wide>.
}
\details{
The wide option enables the confirmed, deaths, and recovered cases to appear in columns instead of in just one column.
}
\examples{
\dontrun{
cases <- get_cases() # standard
}
\dontrun{
cases <- get_cases(wide = TRUE) # wide dataframe
}

}
